+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Pre-release]
 - Update to Web Reader version 4.3.1
 - Add TOC for single resource PDFs
+- Refactored Playwright/Cucumber tests to use new dropdown action function
 
 ## [0.17.1]
 - Replace dgx-header-component and DS Footer with Header/Footer components from nypl-header-app